refactor kernels; backward
