+++ sparta/nn/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from sparta.specializer import OperatorBase, SparseLinear, SparseSoftmax, SparseAttention
+# from sparta.specializer import OperatorBase, SparseLinear, SparseSoftmax, SparseAttention
+from sparta.specializer import OperatorBase
 from sparta.nn.module_tuner import tune_combined_module as tune
