+++ sparta/specializer/kernels/__init__.py
@@ -2,5 +2,4 @@
 # Licensed under the MIT license.
 
 from sparta.specializer.kernels.kernel_base import KernelBase
-from sparta.specializer.kernels.matmul import MatMulKernelBase, SparTATemplateSparseMatMulKernel, OpenAITemplateSparseMatMulKernel
-from sparta.specializer.kernels.softmax import SoftmaxKernelBase, SparTATemplateSparseSoftmaxKernel
+from sparta.specializer.kernels.matmul import SparseMatMulKernel, SparTASparseMatMulKernel, OpenAISparseMatMulKernel
