+++ sparta/specializer/kernels/templates/openai_sparse_matmul_dsd.cuh.j2
@@ -37,7 +37,9 @@ __global__ void BLOCK_SPARSE_MATMUL_32_64_32(
     tiling k dimension
     tile size: 32x64x32
     smm_sd_d_nn: sparse matmul, dense (MxK, along K) x sparse (NxK, along K, K major bcsr) -> dense (MxN, along N)
-    smm_sd_d_nt: sparse matmul, dense (MxK, along K) x sparse (KxN, along N, K major bcsr, need transpose) -> dense (MxN, along N)
+    smm_sd_d_nt: sparse matmul, dense (MxK, along K) x sparse (KxN, along N, K major bcsr, need TRANSPOSE_B) -> dense (MxN, along N)
+    smm_sd_d_tn: sparse matmul, dense (KxM, along M, need TRANSPOSE_A) x sparse (NxK, along K, K major bcsr) -> dense (MxN, along N)
+    smm_sd_d_tt: sparse matmul, dense (KxM, along M, need TRANSPOSE_A) x sparse (KxN, along N, K major bcsr, need TRANSPOSE_B) -> dense (MxN, along N)
     dense matrix X block sparse matrix (block size: 64x32) -> dense matrix
     */
     const int BLOCK_SIZE_M = 32;  // 64
@@ -80,12 +82,18 @@ __global__ void BLOCK_SPARSE_MATMUL_32_64_32(
     assert(THREAD_SIZE_K % 16 == 0);
 
     // A is stored in dense format
+    {% if TRANSPOSE_A %}
+    uint ori_offsetA00 = tid / (BLOCK_SIZE_M/4) * M + by * BLOCK_SIZE_M + (tid % (BLOCK_SIZE_M/4)) * 4;
+    uint ori_offsetA16 = ori_offsetA00 + M * 32;
+    uint storA = (tid * 4 + tid / (BLOCK_SIZE_M/4) / 4 * 2) * 4;
+    {% else %}
     uint ori_offsetA00 = (by * 32 + ty) * K + tx * 4;
     uint ori_offsetA16 = ori_offsetA00 + K * 16;
+    {% endif %}
 
     {% if COMPRESSED %}
     // B is stored in sparse format
-        {% if TRANSPOSE %}
+        {% if TRANSPOSE_B %}
     uint offsetB00 = B_block_ptr[bx] * BLOCK_SIZE_N * BLOCK_SIZE_K + ty * BLOCK_SIZE_K + tx * 4;
     uint offsetB16 = offsetB00 + BLOCK_SIZE_K * 16;
         {% else %}
@@ -95,7 +103,7 @@ __global__ void BLOCK_SPARSE_MATMUL_32_64_32(
         {% endif %}
     {% else %}
     // B is stored in dense format
-        {% if TRANSPOSE %}
+        {% if TRANSPOSE_B %}
     uint ori_offsetB00 = (bx * 32 + ty) * K + tx * 4;
     uint ori_offsetB16 = ori_offsetB00 + K * 16;
         {% else %}
@@ -128,10 +136,15 @@ __global__ void BLOCK_SPARSE_MATMUL_32_64_32(
     int index_start = B_block_ptr[bx], index_end = B_block_ptr[bx+1];
     for (int bcsr_row_idx = index_start; bcsr_row_idx < index_end; bcsr_row_idx += 1)
     {
+        {% if TRANSPOSE_A %}
+        uint offsetA00 = ori_offsetA00 + 64 * B_block_idx[bcsr_row_idx] * M;
+        uint offsetA16 = ori_offsetA16 + 64 * B_block_idx[bcsr_row_idx] * M;
+        {% else %}
         uint offsetA00 = ori_offsetA00 + 64 * B_block_idx[bcsr_row_idx];
         uint offsetA16 = ori_offsetA16 + 64 * B_block_idx[bcsr_row_idx];
+        {% endif %}
         {% if not COMPRESSED %}
-            {% if TRANSPOSE %}
+            {% if TRANSPOSE_B %}
         uint offsetB00 = ori_offsetB00 + 64 * B_block_idx[bcsr_row_idx];
         uint offsetB16 = ori_offsetB16 + 64 * B_block_idx[bcsr_row_idx];
             {% else %}
@@ -152,6 +165,16 @@ __global__ void BLOCK_SPARSE_MATMUL_32_64_32(
 
         __syncthreads();
 
+        {% if TRANSPOSE_A %}
+        *(float*)&bShare[storA + (0*65*32 + 0)*4] = a00.x;
+        *(float*)&bShare[storA + (0*65*32 + 1)*4] = a00.y;
+        *(float*)&bShare[storA + (0*65*32 + 2)*4] = a00.z;
+        *(float*)&bShare[storA + (0*65*32 + 3)*4] = a00.w;
+        *(float*)&bShare[storA + (32*32 + 8*2 + 0*65*32 + 0)*4] = a16.x;
+        *(float*)&bShare[storA + (32*32 + 8*2 + 0*65*32 + 1)*4] = a16.y;
+        *(float*)&bShare[storA + (32*32 + 8*2 + 0*65*32 + 2)*4] = a16.z;
+        *(float*)&bShare[storA + (32*32 + 8*2 + 0*65*32 + 3)*4] = a16.w;
+        {% else %}
         *(float*)&bShare[storAB + (0*32 +  0 + 0*65*32)*4] = a00.x;
         *(float*)&bShare[storAB + (1*32 +  0 + 0*65*32)*4] = a00.y;
         *(float*)&bShare[storAB + (2*32 +  0 + 0*65*32)*4] = a00.z;
@@ -160,8 +183,9 @@ __global__ void BLOCK_SPARSE_MATMUL_32_64_32(
         *(float*)&bShare[storAB + (1*32 + 16 + 0*65*32)*4] = a16.y;
         *(float*)&bShare[storAB + (2*32 + 16 + 0*65*32)*4] = a16.z;
         *(float*)&bShare[storAB + (3*32 + 16 + 0*65*32)*4] = a16.w;
+        {% endif %}
 
-        {% if TRANSPOSE %}
+        {% if TRANSPOSE_B %}
         *(float*)&bShare[storAB + (0*32 +  0 + 1*65*32)*4] = b00.x;
         *(float*)&bShare[storAB + (1*32 +  0 + 1*65*32)*4] = b00.y;
         *(float*)&bShare[storAB + (2*32 +  0 + 1*65*32)*4] = b00.z;
