+++ sparta/specializer/operators/__init__.py
@@ -2,6 +2,3 @@
 # Licensed under the MIT license.
 
 from sparta.specializer.operators.operator_base import OperatorBase
-from sparta.specializer.operators.sparse_linear import SparseLinear
-from sparta.specializer.operators.sparse_softmax import SparseSoftmax
-from sparta.specializer.operators.sparse_attention import SparseAttention
