+++ sparta/specializer_old/kernels/templates/sparta_sparse_matmul_dds.cuh.j2
@@ -13,33 +13,35 @@ const int TN = {{ THREAD_SIZE_N_VALUE }};
 
 __global__ void BLOCK_SPARSE_MATMUL(
     float* input_A,
-    float* input_B_val,
-    int* input_B_block_ptr,
-    int* input_B_block_idx,
+    float* input_B,
     {% if BIASED %}float* input_bias,{% endif %}
-    float* output_C
+    int* output_C_row_idx,
+    int* output_C_col_idx,
+    int* output_C_block_nnz,
+    float* output_C_val
 ) {
     float * A = reinterpret_cast<float*>(input_A);
-    float * B_val = reinterpret_cast<float*>(input_B_val);
-    int * B_block_ptr = reinterpret_cast<int*>(input_B_block_ptr);
-    int * B_block_idx = reinterpret_cast<int*>(input_B_block_idx);
+    float * B = reinterpret_cast<float*>(input_B);
     {% if BIASED %}float * bias = reinterpret_cast<float*>(input_bias);{% endif %}
-    float * C = reinterpret_cast<float*>(output_C);
+    int * C_row_idx = reinterpret_cast<int*>(output_C_row_idx);
+    int * C_col_idx = reinterpret_cast<int*>(output_C_col_idx);
+    int * C_block_nnz = reinterpret_cast<int*>(output_C_block_nnz);
+    float * C_val = reinterpret_cast<float*>(output_C_val);
 
-    int by = blockIdx.y;
-    int bx = blockIdx.x;
+    int bx = C_col_idx[blockIdx.x]; // N
+    int by = C_row_idx[blockIdx.x]; // M
     int ty = threadIdx.y;
     int tx = threadIdx.x;
 
-    A += M * K * blockIdx.z;
+    A += M * K * blockIdx.y;
+    B += K * N * blockIdx.y;
     {% if COMPRESSED %}
-    B_val += B_block_ptr[N / BN] * BN * BK * blockIdx.z;
+    C_val += C_block_nnz[0] * BM * BN * blockIdx.y;
     {% else %}
-    B_val += K * N * blockIdx.z;
+    C_val += M * N * blockIdx.y;
     {% endif %}
-    C += M * N * blockIdx.z;
     {% if BIASED %}
-    bias += N * blockIdx.z;
+    bias += N * blockIdx.y;
     {% endif %}
 
     __shared__ float As[BM * BK];
@@ -49,8 +51,8 @@ __global__ void BLOCK_SPARSE_MATMUL(
     float a_frag[TM][TK];
     float b_frag[TN][TK];
 
-    int A_THREAD_PER_ROW = {% if TRANSPOSE_A %}BM{% else %}BK{% endif %} / 4;
-    int B_THREAD_PER_ROW = {% if TRANSPOSE_B %}BK{% else %}BN{% endif %} / 4;
+    int A_THREAD_PER_ROW = BK / 4;
+    int B_THREAD_PER_ROW = {% if TRANSPOSE %}BK{% else %}BN{% endif %} / 4;
 
     int bszy = BM / TM;
     int bszx = BN / TN;
@@ -68,46 +70,27 @@ __global__ void BLOCK_SPARSE_MATMUL(
     int A_BLOCK_COL_START = tid % A_THREAD_PER_ROW * 4;
     int B_BLOCK_COL_START = tid % B_THREAD_PER_ROW * 4;
 
-    int index_start = B_block_ptr[bx], index_end = B_block_ptr[bx+1];
-
     const int vBLOCK_SIZE_M = BM / TM;
     const int vBLOCK_SIZE_N = BN / TN;
-    {% if TRANSPOSE_A or TRANSPOSE_B %}float4 tmp_float4;{% endif %}
-    for (int tile_block_idx = index_start; tile_block_idx < index_end; tile_block_idx += 1) {
-        int tile_idx = B_block_idx[tile_block_idx] * BK;
+    {% if TRANSPOSE %}float4 tmp_float4;{% endif %}
+    for (int tile_idx = 0; tile_idx < K; tile_idx += BK) {
         #pragma unroll
-        for (int k = 0; k < {% if TRANSPOSE_A %}BK{% else %}BM{% endif %}; k += A_TILE_ROW_STRIDE) {
-            {% if TRANSPOSE_A %}
-            tmp_float4 = (reinterpret_cast<float4*>(&A[(tile_idx + A_BLOCK_ROW_START+k) * M + by*BM + A_BLOCK_COL_START]))[0];
-            As[(A_BLOCK_COL_START+0) * BK + k+A_BLOCK_ROW_START] = tmp_float4.x;
-            As[(A_BLOCK_COL_START+1) * BK + k+A_BLOCK_ROW_START] = tmp_float4.y;
-            As[(A_BLOCK_COL_START+2) * BK + k+A_BLOCK_ROW_START] = tmp_float4.z;
-            As[(A_BLOCK_COL_START+3) * BK + k+A_BLOCK_ROW_START] = tmp_float4.w;
-            {% else %}
+        for (int k = 0; k < BM; k += A_TILE_ROW_STRIDE) {
             *((float4 *)(&As[(k+A_BLOCK_ROW_START) * BK + A_BLOCK_COL_START])) =
-                *((float4 *)(&A[(by*BM + k+A_BLOCK_ROW_START) * K + tile_idx + A_BLOCK_COL_START]));
-            {% endif %}
+                *((float4 *)(&A[(by*BM+k+A_BLOCK_ROW_START) * K + tile_idx+A_BLOCK_COL_START]));
         }
 
         #pragma unroll
-        for (int k = 0; k < {% if TRANSPOSE_B %}BN{% else %}BK{% endif %}; k += B_TILE_ROW_STRIDE) {
-            {% if TRANSPOSE_B %}
-                {% if COMPRESSED %}
-            tmp_float4 = (reinterpret_cast<float4*>(&B_val[tile_block_idx * BN * BK + (k+B_BLOCK_ROW_START) * BK + B_BLOCK_COL_START]))[0];
-                {% else %}
-            tmp_float4 = (reinterpret_cast<float4*>(&B_val[(bx*BN + B_BLOCK_ROW_START+k) * K + tile_idx + B_BLOCK_COL_START]))[0];
-                {% endif %}
+        for (int k = 0; k < {% if TRANSPOSE %}BN{% else %}BK{% endif %}; k += B_TILE_ROW_STRIDE) {
+            {% if TRANSPOSE %}
+            tmp_float4 = (reinterpret_cast<float4*>(&B[(bx*BN + B_BLOCK_ROW_START+k) * K + tile_idx + B_BLOCK_COL_START]))[0];
             Bs[(B_BLOCK_COL_START+0) * BN + k+B_BLOCK_ROW_START] = tmp_float4.x;
             Bs[(B_BLOCK_COL_START+1) * BN + k+B_BLOCK_ROW_START] = tmp_float4.y;
             Bs[(B_BLOCK_COL_START+2) * BN + k+B_BLOCK_ROW_START] = tmp_float4.z;
             Bs[(B_BLOCK_COL_START+3) * BN + k+B_BLOCK_ROW_START] = tmp_float4.w;
             {% else %}
             *((float4 *)(&Bs[(k+B_BLOCK_ROW_START) * BN + B_BLOCK_COL_START])) =
-                {% if COMPRESSED %}
-                *((float4 *)(&B_val[tile_block_idx * BN * BK + (k+B_BLOCK_ROW_START) * BN + B_BLOCK_COL_START]));
-                {% else %}
-                *((float4 *)(&B_val[(tile_idx + B_BLOCK_ROW_START+k) * N + bx*BN + B_BLOCK_COL_START]));
-                {% endif %}
+                *((float4 *)(&B[(tile_idx + B_BLOCK_ROW_START+k) * N + bx*BN + B_BLOCK_COL_START]));
             {% endif %}
         }
 
@@ -157,8 +140,13 @@ __global__ void BLOCK_SPARSE_MATMUL(
     for (int thread_x = 0; thread_x < TN; thread_x++) {
         #pragma unroll
         for (int thread_y = 0; thread_y < TM; thread_y+=1) {
-            C[(BM * by + ty + thread_y * vBLOCK_SIZE_M) * N + BN * bx + tx + thread_x * vBLOCK_SIZE_N] =
+            {% if COMPRESSED %}
+            C_val[BM * BN * blockIdx.x + (ty + thread_y * vBLOCK_SIZE_M) * BN + tx + thread_x * vBLOCK_SIZE_N] =
                 (accum[thread_x][thread_y]){% if BIASED %} + bias_local[thread_x]{% endif %};
+            {% else %}
+            C_val[(BM * by + ty + thread_y * vBLOCK_SIZE_M) * N + BN * bx + tx + thread_x * vBLOCK_SIZE_N] =
+                (accum[thread_x][thread_y]){% if BIASED %} + bias_local[thread_x]{% endif %};
+            {% endif %}
         }
     }
 }
