+++ sparta/specializer_old/kernels/templates/sparta_sparse_matmul_dsd.cuh.j2
@@ -49,8 +49,8 @@ __global__ void BLOCK_SPARSE_MATMUL(
     float a_frag[TM][TK];
     float b_frag[TN][TK];
 
-    int A_THREAD_PER_ROW = {% if TRANSPOSE_A %}BM{% else %}BK{% endif %} / 4;
-    int B_THREAD_PER_ROW = {% if TRANSPOSE_B %}BK{% else %}BN{% endif %} / 4;
+    int A_THREAD_PER_ROW = BK / 4;
+    int B_THREAD_PER_ROW = {% if TRANSPOSE %}BK{% else %}BN{% endif %} / 4;
 
     int bszy = BM / TM;
     int bszx = BN / TN;
@@ -72,26 +72,18 @@ __global__ void BLOCK_SPARSE_MATMUL(
 
     const int vBLOCK_SIZE_M = BM / TM;
     const int vBLOCK_SIZE_N = BN / TN;
-    {% if TRANSPOSE_A or TRANSPOSE_B %}float4 tmp_float4;{% endif %}
+    {% if TRANSPOSE %}float4 tmp_float4;{% endif %}
     for (int tile_block_idx = index_start; tile_block_idx < index_end; tile_block_idx += 1) {
         int tile_idx = B_block_idx[tile_block_idx] * BK;
         #pragma unroll
-        for (int k = 0; k < {% if TRANSPOSE_A %}BK{% else %}BM{% endif %}; k += A_TILE_ROW_STRIDE) {
-            {% if TRANSPOSE_A %}
-            tmp_float4 = (reinterpret_cast<float4*>(&A[(tile_idx + A_BLOCK_ROW_START+k) * M + by*BM + A_BLOCK_COL_START]))[0];
-            As[(A_BLOCK_COL_START+0) * BK + k+A_BLOCK_ROW_START] = tmp_float4.x;
-            As[(A_BLOCK_COL_START+1) * BK + k+A_BLOCK_ROW_START] = tmp_float4.y;
-            As[(A_BLOCK_COL_START+2) * BK + k+A_BLOCK_ROW_START] = tmp_float4.z;
-            As[(A_BLOCK_COL_START+3) * BK + k+A_BLOCK_ROW_START] = tmp_float4.w;
-            {% else %}
+        for (int k = 0; k < BM; k += A_TILE_ROW_STRIDE) {
             *((float4 *)(&As[(k+A_BLOCK_ROW_START) * BK + A_BLOCK_COL_START])) =
-                *((float4 *)(&A[(by*BM + k+A_BLOCK_ROW_START) * K + tile_idx + A_BLOCK_COL_START]));
-            {% endif %}
+                *((float4 *)(&A[(by*BM+k+A_BLOCK_ROW_START) * K + tile_idx+A_BLOCK_COL_START]));
         }
 
         #pragma unroll
-        for (int k = 0; k < {% if TRANSPOSE_B %}BN{% else %}BK{% endif %}; k += B_TILE_ROW_STRIDE) {
-            {% if TRANSPOSE_B %}
+        for (int k = 0; k < {% if TRANSPOSE %}BN{% else %}BK{% endif %}; k += B_TILE_ROW_STRIDE) {
+            {% if TRANSPOSE %}
                 {% if COMPRESSED %}
             tmp_float4 = (reinterpret_cast<float4*>(&B_val[tile_block_idx * BN * BK + (k+B_BLOCK_ROW_START) * BK + B_BLOCK_COL_START]))[0];
                 {% else %}
